% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbl.R
\name{gbl}
\alias{gbl}
\title{lacunarity}
\usage{
gbl(
  init_boxwidth,
  end_boxwidth,
  image,
  obserwin = Frame(image),
  raster_type = c("continuous", "binary")
)
}
\arguments{
\item{init_boxwidth}{The box width for the beginning box put in the corner of a raster surface}

\item{end_boxwidth}{The box width for your identified last box}

\item{image}{The image formate of the raster layer}

\item{obserwin}{The optional observation window. If provided, it will be the intersection of the specified observation
window (obserwin) and the non-NA pixels in the input image (image).}

\item{raster_type}{The selection of the raster type. It can be either "continuous" or "binary"}
}
\value{
A dataframe includes box widths (box_width) and their corresponding lacunarity values.
}
\description{
This function calculates the lacunarity value for both binary and continuous raster layers.This function
was developed from the function gblemp of the R package lacunaritycovariance (Author: Kassel Liam Hingee), which
was extended to apply on continuous raster layers, and the box width will be increased with pixel width.
}
\examples{
\dontrun{library(raster);library(maptools);library(RcppRoll)
study_area <- raster(study_area.tif)
im_area <- as.im.Raster(study_area)
LA <- gbl(30,10000,im_area,raster_type="continuous")
}
}
