% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lac.limit.R
\name{lac.limit}
\alias{lac.limit}
\title{lac.limit}
\usage{
lac.limit(
  data,
  box_sizes,
  lacunarity_values,
  start = list(a = 1, b = -1, c = 1)
)
}
\arguments{
\item{data}{A dataframe includes the boxsizes of the gliding boxes and their corresponding lacuanrity value}

\item{box_sizes}{The increased box_sizes of gliding boxes in lacunarity analysis}

\item{lacunarity_values}{The corresponding lacunarity values of the increased box sizes given by lacunarity analysis}

\item{start}{The start value of the curvature simulation. This function used a power function a * x^b + c to simulate
the lacunarity curvature. Through simulation, the flattened point can be obtained by computing the maximum curvature.
A start = list(a = 1, b = -1, c = 1) was set as default. Users can adjusted it based on their data.}
}
\value{
Maximum carvature point
}
\description{
The function calculates the precise point where the lacunarity curve flattened out.
}
\examples{
 
\dontrun{
lac_ndvi <- read.csv("~path/lac_ndvi.csv")
maxcurv <- lac.limit(lac_ndvi, "distance", "GBL")
}
}
