% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.overall.R
\name{model.overall}
\alias{model.overall}
\title{model.overall}
\usage{
model.overall(data, surv_time, status, covariates, exposure_colnames)
}
\arguments{
\item{data}{A data frame containing the survival time, status,covariates,and exposures delineated by different geographic contexts}

\item{surv_time}{Te column name of survivial time}

\item{status}{The column name of status}

\item{covariates}{A character vector of column names for the fixed covariates}

\item{exposure_colnames}{A character vector of column names for the exposure variables, the column names of exposures should be
the combination of prefixes and buffer sizes, and the same exposures should have the same prefixes, for example, the column names
of greenspace for buffer sizes 100, 200, 300, 400 could be bufnd100, bufnd200,bufnd300,bufnd400; for noise could be bufno100,
bufno200, bufno300, and bufno400.Note, the prefixes can only be alphabets, e.g., bufnd,bufpm,bufno..}
}
\value{
A dataframe includes the the coefficients, posterior possibility, and BIC for each exposure-specific cox model
}
\description{
This function calculates the averaged coefficients (i.e.,beta, p-value, z-value, standard error, hazard ratio,
confidence interval) weighted by posterior possibility
}
\examples{
\dontrun{Take our provided fake mortality data (mortality_exposure.csv) as an example:
data <- read.csv("~path/mortality_exposure.csv")
head(data)
covariates <- colnames(data[,c(3:7)])
exposure_colnames <- colnames(data[,c(8:29)])
overall_model <- model.overall(data,"Survival_time","Status",covariates,exposure_colnames)
}
}
