% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata_documentation02.R
\docType{data}
\name{test_lacunarity}
\alias{test_lacunarity}
\title{Example data for function lac.limit}
\format{
A data frame with 341 rows and 2 columns
}
\usage{
data(test_lacunarity)
}
\description{
This dataset is a fake data which contains the lacunarity analysis results of NDVI given by function
lacunarity(), which included the box widths (30-m increment) and their corresponding lacunarity values.
}
\details{
test_lacunarity

A dataset containing example data for computing the precise flattened point for lacunarity curve
}
\keyword{datasets}
