% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata_documentation01.R
\docType{data}
\name{test_survival}
\alias{test_survival}
\title{Example data for function model.pp and model.overall}
\format{
A data frame with 100000 rows and 29 columns
}
\usage{
data(test_survival)
}
\description{
This dataset is a fake data which contains 100000 objects included their survival time,censoring,
covariates and environmental exposures delineated by different buffer sizes. For instance, the bufND100, bufPM100,
and bufno100 indicate the NDVI, PM2.5, and noise delineated by 100-m circular buffers.
}
\details{
test_survival

A dataset containing example data for bayesian model averaging pooled cox regression (model.pp and model.overall)
}
\keyword{datasets}
